+++ parser_redirect.c
@@ -6,7 +6,7 @@
 /*   By: kaisuzuk <kaisuzuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/23 11:48:27 by kaisuzuk          #+#    #+#             */
-/*   Updated: 2025/11/19 14:15:48 by kaisuzuk         ###   ########.fr       */
+/*   Updated: 2025/12/05 02:13:18 by kaisuzuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,14 +51,13 @@ static t_bool	set_redirect(t_redirect *redir, t_redirect_info info,
 static t_redirect	*make_redirection(t_token_list **token, t_token_error *e)
 {
 	t_redirect				*redirect;
-	int						i;
-	const t_redirect_info	redir_table[] = {
+	unsigned long long int	i;
+	const t_redirect_info redir_table[] = {
 	{TK_LESS, r_input_direction, O_RDONLY},
 	{TK_LESS_LESS, r_reading_until, 0},
 	{TK_GREAT, r_output_direction, O_WRONLY | O_CREAT | O_TRUNC},
 	{TK_GREAT_GREAT, r_appending_to, O_WRONLY | O_CREAT | O_APPEND},
 	};
-
 	redirect = (t_redirect *)xcalloc(sizeof(t_redirect), 1);
 	if (!redirect)
 		return (set_parse_error(ST_ERR_NOMEM, NULL, NULL, e), NULL);
